% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{echarts_cv_builder}
\alias{echarts_cv_builder}
\title{echarts cv builder}
\usage{
echarts_cv_builder(echart_data, stat2plot, lang)
}
\arguments{
\item{echart_data}{data to build the echart}

\item{stat2plot}{stat (bias, mae...) to plot}

\item{lang}{selected lang for translations}
}
\description{
Build the cross validations maps
}
