% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{forestdrought_spain_app}
\alias{forestdrought_spain_app}
\title{App launcher}
\usage{
forestdrought_spain_app()
}
\description{
Function to launch the app (in console, in container...)
}
\details{
This function builds the UI (calling any modules necessary), the server
(calling any modules necessary) and return a \code{ShinyApp()} object.
}
