% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestdroughtSpainApp-package.R
\docType{package}
\name{forestdroughtSpainApp-package}
\alias{forestdroughtSpainApp}
\alias{forestdroughtSpainApp-package}
\title{forestdroughtSpainApp: ForestDrought Shiny App For Continental Spain and Balearic Islands}
\description{
Shiny app package wrapper for continental spain medfate results.
}
\author{
\strong{Maintainer}: Víctor Granda \email{victorgrandagarcia@gmail.com}

}
\keyword{internal}
