% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadOutput.R
\name{mod_downloadOutput}
\alias{mod_downloadOutput}
\alias{mod_download}
\title{mod_downloadOutput and mod_download}
\usage{
mod_downloadOutput(id)

mod_download(input, output, session, user_inputs, ts_outputs, lang)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{user_inputs}{reactiveValues containing the user selected inputs}

\item{ts_outputs}{reactiveValues containing the ts calculated ouputs}

\item{lang}{lang selected}
}
\description{
Shiny module to generate the contents of the download tab
}
