% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapOutput.R
\name{mod_mapOutput}
\alias{mod_mapOutput}
\alias{mod_map}
\title{mod_mapOutput and mod_map}
\usage{
mod_mapOutput(id)

mod_map(input, output, session, arrow_sink, lang)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{arrow_sink}{bucket s3 filesystem}

\item{lang}{lang selected}
}
\description{
Shiny module to generate the map with mabox and mapdeck
}
