% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_translateOutput.R
\name{mod_tab_translateOutput}
\alias{mod_tab_translateOutput}
\alias{mod_tab_translate}
\title{mod_tab_translateOutput and mod_tab_translation}
\usage{
mod_tab_translateOutput(id)

mod_tab_translate(input, output, session, tab_title, lang)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{tab_title}{character with the tab title to translate}

\item{lang}{lang reactive}
}
\description{
A shiny module to translate tab titles
}
