% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_techSpecsOutput.R
\name{mod_techSpecsOutput}
\alias{mod_techSpecsOutput}
\alias{mod_techSpecs}
\title{mod_techSpecsOutput and mod_techSpecs}
\usage{
mod_techSpecsOutput(id)

mod_techSpecs(input, output, session, lang)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}
}
\description{
Shiny module to show the technical specifications text
}
\details{
mod_techSpecs generates the crossvalidations table
}
