% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tsOutput.R
\name{mod_tsOutput}
\alias{mod_tsOutput}
\alias{mod_ts}
\title{mod_tsOutput and mod_ts}
\usage{
mod_tsOutput(id)

mod_ts(input, output, session, arrow_sink, lang)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{arrow_sink}{bucket s3 filesystem}

\item{lang}{lang selected}
}
\description{
Shiny module to show the variables timeseries
}
