% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_userInput.R
\name{mod_user}
\alias{mod_user}
\title{mod_user server function}
\usage{
mod_user(input, output, session, lang)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}
}
\description{
mod_user server function
}
