% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_userInput.R
\name{mod_userInput}
\alias{mod_userInput}
\title{mod_userInput and mod_user}
\usage{
mod_userInput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the user inputs
(variable, dates...)
}
