% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{echarts_ts_formatter}
\alias{echarts_ts_formatter}
\title{echarts ts formatter}
\usage{
echarts_ts_formatter(echart)
}
\arguments{
\item{echart}{echarts4r object to format}
}
\description{
Apply the common format (legend, tooltip, theme...) to timeseries (echarts)
}
